# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: API de Gestion des Artistes
  description: |
    API REST permettant la gestion des artistes dans le systeme.
    Cette API permet de creer, lire, mettre a jour et supprimer des artistes.
  version: 1.0.0
  contact:
    name: Support API
    email: support@example.com

servers:
  - url: http://localhost:8080
    description: Serveur de developpement local

tags:
  - name: artistes
    description: Operations liees a la gestion des artistes

paths:
  /api/artistes:
    get:
      tags:
        - artistes
      summary: Recupere tous les artistes
      description: Retourne la liste complete des artistes enregistres dans le systeme
      operationId: getAllArtists
      responses:
        '200':
          description: Liste des artistes recuperee avec succes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ArtistDto'
        '500':
          description: Erreur serveur interne

    post:
      tags:
        - artistes
      summary: Cree un nouvel artiste
      description: Ajoute un nouvel artiste dans le systeme
      operationId: createArtist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArtistDto'
      responses:
        '200':
          description: Artiste cree avec succes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtistDto'
        '400':
          description: Donnees invalides
        '500':
          description: Erreur serveur interne

  /api/artistes/{id}:
    get:
      tags:
        - artistes
      summary: Recupere un artiste par son ID
      description: Retourne les details d'un artiste specifique
      operationId: getArtistById
      parameters:
        - name: id
          in: path
          description: ID de l'artiste a recuperer
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Artiste trouve
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtistDto'
        '404':
          description: Artiste non trouve
        '500':
          description: Erreur serveur interne

    put:
      tags:
        - artistes
      summary: Met a jour un artiste
      description: Met a jour un artiste du systeme par son ID
      operationId: updateArtist
      parameters:
        - name: id
          in: path
          description: ID de l'artiste a mettre a jour
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArtistDto'
      responses:
        '200':
          description: Artiste mis a jour avec succes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtistDto'
        '400':
          description: Donnees invalides
        '404':
          description: Artiste non trouve
        '500':
          description: Erreur serveur interne

    delete:
      tags:
        - artistes
      summary: Supprime un artiste
      description: Supprime un artiste du systeme par son ID
      operationId: deleteArtist
      parameters:
        - name: id
          in: path
          description: ID de l'artiste a supprimer
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Artiste supprime avec succes
        '404':
          description: Artiste non trouve
        '500':
          description: Erreur serveur interne

components:
  schemas:
    ArtistDto:
      type: object
      description: Representation d'un artiste dans le systeme
      properties:
        id:
          type: integer
          format: int64
          description: Identifiant unique de l'artiste
          example: 1
        nom:
          type: string
          description: Nom de famille de l'artiste
          example: "Dupont"
        prenom:
          type: string
          description: Prenom de l'artiste
          example: "Jean"
        age:
          type: integer
          format: int32
          description: Age de l'artiste
          example: 35
      required:
        - nom
        - prenom
        - age